+++ packages/core/src/reporters/DefaultReporter.ts
@@ -57,7 +57,7 @@ export default class DefaultReporter extends Reporter<Line> {
    * Return the task title with additional metadata.
    */
   // eslint-disable-next-line complexity
-  getLineTitle(task: Task<any> | Routine<any, any>, usedColumns: number = 0): string {
+  getLineTitle(task: Task<any> | Routine<any, any>): { title: string; status: string } {
     const outputLevel = this.tool.config.output;
     // @ts-ignore
     const { tasks = [], routines = [] } = task;
@@ -74,18 +74,21 @@ export default class DefaultReporter extends Reporter<Line> {
       status.push(`${this.calculateTaskCompletion(routines)}/${routines.length}`);
     }
 
-    if (task instanceof Routine && !task.isSkipped() && outputLevel >= 2) {
-      status.push(this.getElapsedTime(task.startTime, task.stopTime));
-    } else if (task instanceof Task && task.statusText) {
-      title = this.style(this.strip(task.statusText), 'pending');
+    if (task instanceof Routine) {
+      if (!task.isSkipped() && outputLevel >= 2) {
+        status.push(this.getElapsedTime(task.startTime, task.stopTime));
+      }
+    } else if (task instanceof Task) {
+      title = this.style(this.console.strip(task.statusText), 'pending');
     }
 
-    // eslint-disable-next-line no-magic-numbers
-    const columns = process.stdout.columns || 80;
-    const fullStatus =
-      status.length > 0 && outputLevel >= 1 ? this.style(` [${status.join(', ')}]`, 'pending') : '';
-
-    return this.truncate(title, columns - usedColumns - fullStatus.length) + fullStatus;
+    return {
+      status:
+        status.length > 0 && outputLevel >= 1
+          ? this.style(` [${status.join(', ')}]`, 'pending')
+          : '',
+      title,
+    };
   }
 
   handleStart = (routines: Routine<any, any>[]) => {
@@ -174,13 +177,13 @@ export default class DefaultReporter extends Reporter<Line> {
     }
 
     // Title
-    const usedColumns = indent + key.length;
+    const { title, status } = this.getLineTitle(routine);
 
-    if (task) {
-      output += this.style(this.getLineTitle(task, usedColumns), 'pending');
-    } else {
-      output += this.getLineTitle(routine, usedColumns);
-    }
+    output +=
+      this.console.truncate(
+        task ? this.style(title, 'pending') : title,
+        this.console.size().columns - output.length - status.length,
+      ) + status;
 
     this.console.write(output, 1);
   }